refactor(types): move implementation from types to kernel
