+++ packages/@textlint/kernel/package.json
@@ -46,6 +46,7 @@
     "structured-source": "^3.0.2"
   },
   "devDependencies": {
+    "@types/structured-source": "^3.0.0",
     "@textlint/markdown-to-ast": "^6.1.3",
     "@types/deep-equal": "^1.0.1",
     "@types/mocha": "^5.2.6",
