+++ packages/@textlint/kernel/src/task/linter-task.ts
@@ -3,8 +3,10 @@
 import CoreTask from "./textlint-core-task";
 import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface";
 import { TextlintFilterRuleDescriptors, TextlintRuleDescriptors } from "../descriptor";
-import { TextlintFilterRuleContext, TextlintRuleContext, TextlintSourceCode } from "@textlint/types";
+import { TextlintSourceCode } from "@textlint/types";
 import { getSeverity } from "../shared/rule-severity";
+import { TextlintRuleContextImpl } from "../context/TextlintRuleContextImpl";
+import { TextlintFilterRuleContextImpl } from "../context/TextlintFilterRuleContextImpl";
 
 export interface TextLintCoreTaskArgs {
     config: TextlintKernelConstructorOptions;
@@ -51,7 +53,7 @@ export default class TextLintCoreTask extends CoreTask {
         // see https://github.com/textlint/textlint/issues/219
         this.ruleDescriptors.lintableDescriptors.forEach(ruleDescriptor => {
             const ruleOptions = ruleDescriptor.normalizedOptions;
-            const ruleContext = new TextlintRuleContext({
+            const ruleContext = new TextlintRuleContextImpl({
                 ruleId: ruleDescriptor.id,
                 severityLevel: getSeverity(ruleOptions),
                 sourceCode,
@@ -62,7 +64,7 @@ export default class TextLintCoreTask extends CoreTask {
         });
         // setup "filters" field
         this.filterRuleDescriptors.descriptors.forEach(filterDescriptor => {
-            const ruleContext = new TextlintFilterRuleContext({
+            const ruleContext = new TextlintFilterRuleContextImpl({
                 ruleId: filterDescriptor.id,
                 sourceCode,
                 ignoreReport,
