+++ packages/@textlint/kernel/test/util/dummy-source-code.js
@@ -2,12 +2,12 @@
 "use strict";
 import { parse } from "@textlint/markdown-to-ast";
 import * as path from "path";
-import { TextlintSourceCode } from "@textlint/types";
+import { TextlintSourceCodeImpl } from "../../src/context/TextlintSourceCodeImpl";
 
 const testPath = path.join(__dirname, "fixtures", "test.md");
 const dummyCode = `dummyCode`;
 export default function createDummySourceCode(code = dummyCode, filePath = testPath) {
-    return new TextlintSourceCode({
+    return new TextlintSourceCodeImpl({
         code,
         ast: parse(code),
         ext: path.extname(filePath),
