+++ packages/@textlint/types/package.json
@@ -42,14 +42,12 @@
     "tabWidth": 4
   },
   "dependencies": {
-    "@textlint/ast-node-types": "^4.2.2",
-    "structured-source": "^3.0.2"
+    "@textlint/ast-node-types": "^4.2.2"
   },
   "devDependencies": {
     "@textlint/markdown-to-ast": "^6.1.3",
     "@types/mocha": "^5.2.6",
     "@types/node": "^11.13.8",
-    "@types/structured-source": "^3.0.0",
     "cross-env": "^5.2.0",
     "mocha": "^6.1.4",
     "prettier": "^1.17.0",
